+++ actionpack/lib/action_dispatch/journey/router.rb
@@ -47,6 +47,7 @@ module ActionDispatch
 
           req.path_parameters = tmp_params
           req.route_uri_pattern = route.path.spec.to_s
+          req.current_route = route
 
           _, headers, _ = response = route.app.serve(req)