+++ actionpack/CHANGELOG.md
@@ -75,6 +75,11 @@
 
     *Jon Dufresne*
 
+*   Prioritize the current route when multiple recall matches are possible with
+    `ActionDispatch::Routing::UrlFor`.
+
+    *Joé Dupuis*
+
 *   `config.action_dispatch.cookies_serializer` now accepts `:message_pack` and
     `:message_pack_allow_marshal` as serializers. These serializers require the
     [`msgpack` gem](https://rubygems.org/gems/msgpack) (>= 1.7.0).